[Fix] 1:1문의 프론트 수정 #58
