+++ src/app/qna/page.module.css
@@ -3,21 +3,7 @@
 	margin-top: 50px;
 }
 
-.qnaAside{
-	width: 130px;
-}
 
-.qnaNotice{
-	font-weight: bold;
-	font-size: 20px;
-
-}
-
-.qnaAside li{
-	margin-top: 15px;
-	list-style: none;
-	font-size: 15px;
-}
 
 
 .qnaSection{
