+++ src/components/etc/etcsidebar.module.css
@@ -1,0 +1,20 @@
+
+
+
+.etcAside{
+	width: 130px;
+}
+
+
+
+.etcNotice{
+	font-weight: bold;
+	font-size: 20px;
+
+}
+
+.etcAsideLi{
+	margin-top: 15px;
+	list-style: none;
+	font-size: 15px;
+}
